requirements.txt
matplotlib>=3.4.3
mypy-extensions>=0.4.3
numpy>=1.21.2
pandas>=1.3.2
pyyaml>=5.1
seaborn>=0.10.1
scikit-learn>=1.0.2
scipy>=1.7.3
torch>=1.9.0
tqdm>=4.62.2
pyieoe>=0.1.1
pingouin>=0.4.0